Rename project from resolver to container
