+++ README.md
@@ -1,19 +1,19 @@
 # Gacela Resolver
 
 <p align="center">
-  <a href="https://github.com/gacela-project/resolver/actions">
-    <img src="https://github.com/gacela-project/resolver/workflows/CI/badge.svg" alt="GitHub Build Status">
+  <a href="https://github.com/gacela-project/container/actions">
+    <img src="https://github.com/gacela-project/container/workflows/CI/badge.svg" alt="GitHub Build Status">
   </a>
-  <a href="https://scrutinizer-ci.com/g/gacela-project/resolver/?branch=main">
-    <img src="https://scrutinizer-ci.com/g/gacela-project/resolver/badges/quality-score.png?b=main" alt="Scrutinizer Code Quality">
+  <a href="https://scrutinizer-ci.com/g/gacela-project/container/?branch=main">
+    <img src="https://scrutinizer-ci.com/g/gacela-project/container/badges/quality-score.png?b=main" alt="Scrutinizer Code Quality">
   </a>
-  <a href="https://scrutinizer-ci.com/g/gacela-project/resolver/?branch=main">
-    <img src="https://scrutinizer-ci.com/g/gacela-project/resolver/badges/coverage.png?b=main" alt="Scrutinizer Code Coverage">
+  <a href="https://scrutinizer-ci.com/g/gacela-project/container/?branch=main">
+    <img src="https://scrutinizer-ci.com/g/gacela-project/container/badges/coverage.png?b=main" alt="Scrutinizer Code Coverage">
   </a>
-  <a href="https://shepherd.dev/github/gacela-project/resolver">
-    <img src="https://shepherd.dev/github/gacela-project/resolver/coverage.svg" alt="Psalm Type-coverage Status">
+  <a href="https://shepherd.dev/github/gacela-project/container">
+    <img src="https://shepherd.dev/github/gacela-project/container/coverage.svg" alt="Psalm Type-coverage Status">
   </a>
-  <a href="https://github.com/gacela-project/resolver/blob/master/LICENSE">
+  <a href="https://github.com/gacela-project/container/blob/master/LICENSE">
     <img src="https://img.shields.io/badge/License-MIT-green.svg" alt="MIT Software License">
   </a>
 </p>
@@ -21,7 +21,7 @@
 ## Installation
 
 ```bash
-composer require gacela-project/resolver
+composer require gacela-project/container
 ```
 
 ## Usage
