+++ tests/Unit/DependencyResolverTest.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace GacelaTest\Unit;
 
-use Gacela\Resolver\DependencyResolver;
-use Gacela\Resolver\Exception\DependencyInvalidArgumentException;
-use Gacela\Resolver\Exception\DependencyNotFoundException;
+use Gacela\Container\DependencyResolver;
+use Gacela\Container\Exception\DependencyInvalidArgumentException;
+use Gacela\Container\Exception\DependencyNotFoundException;
 use GacelaTest\Fake\ClassWithInterfaceDependencies;
 use GacelaTest\Fake\ClassWithObjectDependencies;
 use GacelaTest\Fake\ClassWithoutDependencies;
